chore(test): setup basic test
- install mocha, chai, chai-http
- refactor index.js to add app.js
- add base route integration test
